+++ dzVents/runtime/tests/testTime.lua
@@ -1243,96 +1243,96 @@ describe('Time', function()
 				end)
 			end)
 
-			describe('at sunAtSouth', function()
+			describe('at solarnoon', function()
 
-				it('should return true if it is at sunAtSsouth', function()
+				it('should return true if it is at solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 664 }
 					local t = Time('2017-01-01 11:04:00')
-					assert.is_true(t.ruleIsAtSunAtSouth('at sunatsouth'))
+					assert.is_true(t.ruleIsAtSolarnoon('at solarnoon'))
 				end)
 
-				it('should return if it is not at sunAtSouth', function()
+				it('should return if it is not at solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 63 }
 					local t = Time('2017-01-01 01:04:00')
-					assert.is_false(t.ruleIsAtSunAtSouth('at sunatsouth'))
+					assert.is_false(t.ruleIsAtSolarnoon('at solarnoon'))
 				end)
 
 				it('should return nil if the rule is not present', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:04:00')
-					assert.is_nil(t.ruleIsAtSunAtSouth('at blabalbba'))
+					assert.is_nil(t.ruleIsAtSolarnoon('at blabalbba'))
 				end)
 
 				it('should detect the rule within a random string', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:04:00')
-					assert.is_true(t.ruleIsAtSunAtSouth('some random at sunatsouth text'))
+					assert.is_true(t.ruleIsAtSolarnoon('some random at solarnoon text'))
 				end)
 			end)
 
-			describe('xx minutes before sunAtSouth', function()
+			describe('xx minutes before solarnoon', function()
 
-				it('should return true if it is xx minutes before sunAtSouth', function()
+				it('should return true if it is xx minutes before solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:02:00')
-					assert.is_true(t.ruleIsBeforeSunAtSouth('2 minutes before sunatsouth'))
+					assert.is_true(t.ruleIsBeforeSolarnoon('2 minutes before solarnoon'))
 				end)
 
-				it('should return if it is more than 2 minutes before sunAtSouth', function()
+				it('should return if it is more than 2 minutes before solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:01:00')
-					assert.is_false(t.ruleIsBeforeSunAtSouth('2 minutes before sunatsouth'))
+					assert.is_false(t.ruleIsBeforeSolarnoon('2 minutes before solarnoon'))
 				end)
 
-				it('should return if it is less than 2 minutes before sunAtSouth', function()
+				it('should return if it is less than 2 minutes before solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:03:00')
-					assert.is_false(t.ruleIsBeforeSunAtSouth('2 minutes before sunatsouth'))
+					assert.is_false(t.ruleIsBeforeSolarnoon('2 minutes before solarnoon'))
 				end)
 
 				it('should return nil if the rule is not present', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:04:00')
-					assert.is_nil(t.ruleIsBeforeSunAtSouth('minutes before sunatsouth'))
+					assert.is_nil(t.ruleIsBeforeSolarnoon('minutes before solarnoon'))
 				end)
 
 				it('should detect the rule within a random string', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:02:00')
-					assert.is_true(t.ruleIsBeforeSunAtSouth('some random 2 minutes before sunatsouth text'))
+					assert.is_true(t.ruleIsBeforeSolarnoon('some random 2 minutes before solarnoon text'))
 				end)
 			end)
 
-			describe('xx minutes after sunAtSouth', function()
+			describe('xx minutes after solarnoon', function()
 
-				it('should return true if it is xx minutes after sunAtSouth', function()
+				it('should return true if it is xx minutes after solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:06:00')
-					assert.is_true(t.ruleIsAfterSunAtSouth('2 minutes after sunatsouth'))
+					assert.is_true(t.ruleIsAfterSolarnoon('2 minutes after solarnoon'))
 				end)
 
-				it('should return if it is more less 2 minutes after sunAtSouth', function()
+				it('should return if it is more less 2 minutes after solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:05:00')
-					assert.is_false(t.ruleIsAfterSunAtSouth('2 minutes after sunatsouth'))
+					assert.is_false(t.ruleIsAfterSolarnoon('2 minutes after solarnoon'))
 				end)
 
-				it('should return if it is more than 2 minutes after sunAtSouth', function()
+				it('should return if it is more than 2 minutes after solarnoon', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:07:00')
-					assert.is_false(t.ruleIsAfterSunAtSouth('2 minutes after sunatsouth'))
+					assert.is_false(t.ruleIsAfterSolarnoon('2 minutes after solarnoon'))
 				end)
 
 				it('should return nil if the rule is not present', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:04:00')
-					assert.is_nil(t.ruleIsAfterSunAtSouth('minutes after sunatsouth'))
+					assert.is_nil(t.ruleIsAfterSolarnoon('minutes after solarnoon'))
 				end)
 
 				it('should detect the rule within a random string', function()
 					_G.timeofday = { ['SunAtSouthInMinutes'] = 64 }
 					local t = Time('2017-01-01 01:06:00')
-					assert.is_true(t.ruleIsAfterSunAtSouth('some random 2 minutes after sunatsouth text'))
+					assert.is_true(t.ruleIsAfterSolarnoon('some random 2 minutes after solarnoon text'))
 				end)
 			end)
 
@@ -1457,7 +1457,7 @@ describe('Time', function()
 						assert.is_false(t.ruleMatchesBetweenRange('between sunset and sunrise'))
 					end)
 
-					it('between sunAtSouth and sunrise', function()
+					it('between solarnoon and sunrise', function()
 						_G.timeofday = {
 							['SunriseInMinutes'] = 360 , -- 06:00
 							['SunsetInMinutes'] = 1080 ,
@@ -1466,10 +1466,10 @@ describe('Time', function()
 
 						-- time between 07:00 and 12:00
 						local t = Time('2017-01-01 07:04:00')
-						assert.is_true(t.ruleMatchesBetweenRange('between sunrise and sunatsouth'))
+						assert.is_true(t.ruleMatchesBetweenRange('between sunrise and solarnoon'))
 
 						t = Time('2017-01-01 17:00:00')
-						assert.is_false(t.ruleMatchesBetweenRange('between sunrise and sunatsouth'))
+						assert.is_false(t.ruleMatchesBetweenRange('between sunrise and solarnoon'))
 					end)
 
 					it('every x minute between sunset and sunrise', function()